refactor(cancel-button): move CancelButtonComponent out of Pathogen module
