+++ app/views/groups/bots/personal_access_tokens/_generate_personal_access_token_modal.html.erb
@@ -29,7 +29,7 @@
     </div>
 
     <div class="mt-4 flex gap-2">
-      <%= render Pathogen::CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
+      <%= render CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
       <%= form.submit t('common.controls.submit'), tone: :primary, emphasis: :solid %>
     </div>
   <% end %>
