+++ app/views/groups/group_links/_form.html.erb
@@ -130,7 +130,7 @@ if Flipper.enabled?(:v2_select2, Current.user)
     </div>
 
     <div class="flex justify-start gap-1">
-      <%= render Pathogen::CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
+      <%= render CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
 
       <%= form.submit I18n.t('common.actions.share'),
                     tone: :primary,
