+++ app/views/groups/new_subgroup.html.erb
@@ -19,7 +19,7 @@
       } %>
 
       <div class="flex gap-2">
-        <%= render Pathogen::CancelButtonComponent.new(href: group_path(@group)) %>
+        <%= render CancelButtonComponent.new(href: group_path(@group)) %>
 
         <%= form.submit t(:"groups.new_subgroup.submit"),
                       tone: :primary,
