+++ app/views/projects/bots/_bot_modal.html.erb
@@ -91,7 +91,7 @@
     <% end %>
 
     <div class="mt-4 flex gap-2">
-      <%= render Pathogen::CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
+      <%= render CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
       <%= form.submit t('common.controls.submit'), tone: :primary, emphasis: :solid %>
     </div>
   <% end %>
