+++ app/views/projects/members/_edit_modal.html.erb
@@ -82,7 +82,7 @@
       </div>
 
       <div class="flex justify-start gap-1">
-        <%= render Pathogen::CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
+        <%= render CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
 
         <%= form.submit t(:"common.actions.save"),
                     tone: :primary,
