+++ app/views/projects/samples/edit.html.erb
@@ -20,7 +20,7 @@
           <%= render partial: "form_fields", locals: { form: form, sample: @sample } %>
 
           <div class="flex gap-2">
-            <%= render Pathogen::CancelButtonComponent.new(href: namespace_project_sample_path(id: @sample.id)) %>
+            <%= render CancelButtonComponent.new(href: namespace_project_sample_path(id: @sample.id)) %>
 
             <%= form.submit t("projects.samples.edit.submit_button"),
                           tone: :primary,
