+++ app/views/projects/workflow_executions/_edit_dialog.html.erb
@@ -41,7 +41,7 @@
     </div>
 
     <div class="mt-4 flex gap-2">
-      <%= render Pathogen::CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
+      <%= render CancelButtonComponent.new(data: { action: "click->viral--dialog#close" }) %>
 
       <%= form.submit t("projects.workflow_executions.edit_dialog.submit_button"),
                   tone: :primary,
